feat(About): add checkUpdateAtStart setting
